+++ R/write.R
@@ -7,6 +7,7 @@
 #'   function is applied to each element of the list.  The default `"auto"`
 #'   uses `toJSON()` if the package `jsonlite` is available, otherwise
 #'
+#'
 #' @param x An object to write to file
 #' @param path The file or connection to write to (dependent on part by method)
 #' @param method The method of saving the file.  When `default`, the method is
@@ -135,20 +136,22 @@ write_file_md5 <- function(
 #' @export
 #' @rdname write_file_md5
 mark_write_methods <- function() {
-  list(
+  list0(
     "default",
     "csv",
     "csv2",
     "csv3",
     "dcf",
+    "feather",
     "json",
     lines = c("lines", "md", "txt", "qmd", "rmd"),
+    "parquet",
     "rds",
     table = c("table", "delim"),
     "tsv",
     "tsv2",
     "write",
-    yaml = c("yaml", "yml")
+    yaml = c("yaml", "yml"),
   )
 }
 
@@ -250,28 +253,11 @@ mark_write_table <- function(
     col.names <- TRUE
   }
 
-  if (identical(list_hook, "auto") || isTRUE(list_hook)) {
-    if (package_available("jsonlite")) {
-      list_hook <- "mark_to_json"
-    } else {
-      list_hook <- function(x) collapse(shQuote(x, "sh"), sep = ",")
-    }
-  } else if (isFALSE(list_hook)) {
-    list_hook <- function(x) NA_character_
-  } else if (isNA(list_hook)) {
-    list_hook <- function(x) {
-      stop(new_condition(
-        "options(mark.list.hook) is NA but list columns detected",
-        class = "writeFileMd5ListHook"
-      ))
-    }
-  }
-
-  if (!isFALSE(list_hook) && !is.null(list_hook)) {
-    list_hook <- match.fun(list_hook)
-    ok <- vap_lgl(x, is.list)
-    if (any(ok)) {
-      x[ok] <- lapply(x[ok], function(i) vap_chr(i, list_hook))
+  list_hook <- get_list_hook(list_hook)
+  if (!is.null(list_hook)) {
+    cols <- which(vap_lgl(x, is.list))
+    for (col in cols) {
+      x[[col]] <- vap_chr(x[[col]], list_hook)
     }
   }
 
@@ -287,6 +273,39 @@ mark_write_table <- function(
     row.names = row.names,
     col.names = col.names,
     qmethod = qmethod
+  )
+}
+
+get_list_hook <- function(hook) {
+  if (is.function(hook)) {
+    return(hook)
+  }
+
+  if (isTRUE(hook)) {
+    hook <- "auto"
+  } else if (isFALSE(hook)) {
+    hook <- "false"
+  } else if (is.na(hook)) {
+    hook <- "na"
+  }
+
+  switch(
+    hook,
+    auto = if (package_available("jsonlite")) {
+      get_list_hook("json")
+    } else {
+      get_list_hook("default")
+    },
+    json = mark_to_json,
+    default = function(x) collapse(shQuote(x, "sh"), sep = ","),
+    false = function(x) NA_character_,
+    none = NULL,
+    # nolint next: brace_linter.
+    na = function(x) stop(new_condition(
+      "options(mark.list.hook) is NA but list columns detected",
+      class = "writeFileMd5ListHook"
+    )),
+    match.fun(hook)
   )
 }
 
@@ -351,6 +370,57 @@ mark_write_json <- function(x, con) {
   require_namespace("yaml")
   string <- mark_to_json(x)
   mark_write_lines(string, con)
+}
+
+mark_write_feather <- function(x, con, ...) {
+  mark_write_arrow(x, con, ..., .method = "feather")
+}
+
+mark_write_parquet <- function(x, con, ...) {
+  mark_write_arrow(x, con, ..., .method = "parquet")
+}
+
+mark_write_arrow <- function(
+    x,
+    con,
+    ...,
+    .method = c("feather", "parquet")
+) {
+  require_namespace("arrow")
+  .method <- mark::match_param(.method)
+
+  switch(
+    .method,
+    feather = {
+      read <- arrow::read_feather
+      write <- arrow::write_feather
+      clean <- function() NULL
+    },
+    parquet = {
+      read <- arrow::read_parquet
+      write <- arrow::write_parquet
+      clean <- base::gc
+    }
+  )
+
+  if (identical(con, stdout())) {
+    temp <- tempfile()
+    con <- file(temp, open = "wb", encoding = "UTF-8")
+    on.exit({
+      co <- utils::capture.output(read(temp, as_data_frame = FALSE))
+      # Something weird was happening after reading the parquet object on
+      # windows; fs::file_delete() was throwing an EPERM error but file.remove()
+      # wasn't. Adding an explicit gc() seems to do the trick...
+      clean()
+      co <- grep("See $metadata", co, value = TRUE, invert = TRUE, fixed = TRUE)
+      co <- co[nzchar(co)]
+      writeLines(co)
+      safe_close(con)
+      safe_fs_delete(temp)
+    })
+  }
+
+  write(x, sink = con, ...)
 }
 
 # helpers -----------------------------------------------------------------