+++ tests/testthat/test-write.R
@@ -1,5 +1,5 @@
 test_that("write_file_md5() works", {
-  df <- data.frame(a = 1, b = 2)
+  df <- quick_dfl(a = 1, b = 2)
   temp <- withr::local_tempfile()
   expect_output(write_file_md5(df))
   expect_message(write_file_md5(df, temp), NA)
@@ -15,25 +15,36 @@ test_that("write_file_md5() types", {
   foo <- function(method) {
     file <- tempfile()
     on.exit(file.remove(file))
-    df <- data.frame(a = 1, b = "n", c = TRUE)
-    expect_message(write_file_md5(df, file, method = method), NA)
+    x <-
+      if (method %in% c(mark_write_methods()$lines, "write")) {
+        letters
+      } else {
+        quick_dfl(a = 1, b = "n", c = TRUE)
+      }
+      expect_message(
+        write_file_md5(x, file, method = !!method),
+        NA
+      )
   }
 
-  foo("csv")
-  foo("csv2")
-  foo("csv3")
-  foo("dcf")
-  foo("json")
-  foo("rds")
-  foo("table")
-  foo("tsv")
-  foo("tsv2")
-  foo("json")
-  foo("yaml")
+  for (method in unlist0(mark_write_methods())) {
+    foo(method)
+  }
+})
+
+test_that("path warning", {
+  t <- tempfile()
+  on.exit(fs::file_delete(t))
+  x <- structure(quick_dfl(a = 1), path = t)
+  expect_warning(
+    write_file_md5(x, t),
+    "attr(x, \"path\") is being overwritten",
+    fixed = TRUE
+  )
 })
 
 test_that("write_file_md5() errors", {
-  df <- data.frame(a = 1)
+  df <- quick_dfl(a = 1)
   expect_error(
     write_file_md5(df, method = "foo"),
     class = "matchParamMatchError"
@@ -44,7 +55,7 @@ test_that("compression works", {
   foo <- function(ext = "") {
     file <- tempfile(fileext = ext)
     on.exit(unlink(file, recursive = TRUE))
-    df <- data.frame(a = 1)
+    df <- quick_dfl(a = 1)
     write_file_md5(df, file)
   }
 
@@ -54,6 +65,12 @@ test_that("compression works", {
   expect_s3_class(foo(".dcf.xz"), "data.frame")
   expect_s3_class(foo(".csv.zip"), "data.frame")
   expect_s3_class(foo(".tar.gz"), "data.frame")
+
+  expect_error(
+    compress("foo.tar.zip"),
+    "'zip' is not a valid method",
+    fixed = TRUE
+  )
 })
 
 test_that("list columns", {
@@ -64,16 +81,32 @@ test_that("list columns", {
     op <- options(mark.list.hook = method)
     on.exit(options(op))
 
-    df <- data.frame(x = c("a", "b"))
+    df <- quick_dfl(x = c("a", "b"))
     df$y <- list(1:2, 2)
 
+    write_file_md5(df, temp)
+    fs::file_delete(temp)
+
+    df$z <- list(1, 2:3)
     write_file_md5(df, temp)
   }
 
   expect_s3_class(foo("auto"), "data.frame")
+  expect_s3_class(foo("default"), "data.frame")
+  expect_s3_class(foo("json"), "data.frame")
   expect_s3_class(foo(toString), "data.frame")
   expect_s3_class(foo(TRUE), "data.frame")
-  expect_s3_class(foo(NULL), "data.frame")
   expect_s3_class(foo(FALSE), "data.frame")
-  expect_error(foo(NA))
+  expect_s3_class(foo("toString"), "data.frame")
+  expect_error(
+    foo("none"),
+    "unimplemented type 'list' in 'EncodeElement'",
+    fixed = TRUE
+  )
+  expect_error(foo(NA), class = "writeFileMd5ListHookError")
+})
+
+test_that("arrow prints something to stdout()", {
+  expect_snapshot(write_file_md5(quick_dfl(a = 1), method = "feather"))
+  expect_snapshot(write_file_md5(quick_dfl(a = 1), method = "parquet"))
 })